Build the snap using a tar archive produced by python setup.py sdist.
When the semver Python plugin runs inside of the snapcraft build root, it always seems to get the incorrect version for the CLI. For example, if tag 0.17.2 is checked out, the semver tags the package as 0.17.3.dev0.xxx.ddmmyyyy. Seb and I tried to debug this, but couldn't get this done. As a result, Seb recommended to use the sdist archive as the source of the snap build, which worked well.